requirements.txt
numpy==1.18.5
matplotlib==3.3.4
tensorflow-datasets==3.2.1
tensorflow-gpu==2.3.0
tensorflow-probability==0.11.0
tensorflow-addons==0.13.0
